+++ plugin-locationcomponent/src/main/java/com/mapbox/maps/plugin/locationcomponent/LocationComponentExt.kt
@@ -5,6 +5,7 @@ package com.mapbox.maps.plugin.locationcomponent
 import android.content.Context
 import androidx.annotation.DrawableRes
 import androidx.core.content.res.ResourcesCompat
+import com.mapbox.maps.ImageHolder
 import com.mapbox.maps.plugin.LocationPuck2D
 import com.mapbox.maps.plugin.Plugin
 import com.mapbox.maps.plugin.delegates.MapPluginProviderDelegate
@@ -14,13 +15,6 @@ import com.mapbox.maps.plugin.delegates.MapPluginProviderDelegate
  */
 val MapPluginProviderDelegate.location: LocationComponentPlugin
   @JvmName("getLocationComponent")
-  get() = this.getPlugin(Plugin.MAPBOX_LOCATION_COMPONENT_PLUGIN_ID)!!
-
-/**
- * Extension val to get the LocationComponentPlugin instance with interface LocationComponentPlugin2 to handle accuracy ring.
- */
-val MapPluginProviderDelegate.location2: LocationComponentPlugin2
-  @JvmName("getLocationComponent2")
   get() = this.getPlugin(Plugin.MAPBOX_LOCATION_COMPONENT_PLUGIN_ID)!!
 
 private fun Context.getCompatDrawable(@DrawableRes resId: Int) = ResourcesCompat.getDrawable(
@@ -39,17 +33,13 @@ fun LocationComponentPlugin.createDefault2DPuck(
   context: Context,
   withBearing: Boolean = false
 ): LocationPuck2D = LocationPuck2D(
-  topImage = context.getCompatDrawable(R.drawable.mapbox_user_icon),
-  bearingImage = context.getCompatDrawable(
-    if (withBearing)
-      R.drawable.mapbox_user_bearing_icon
-    else
-      R.drawable.mapbox_user_stroke_icon
-  ),
-  shadowImage = context.getCompatDrawable(
-    if (withBearing)
-      R.drawable.mapbox_user_stroke_icon
-    else
-      R.drawable.mapbox_user_icon_shadow
-  )
+  topImage = ImageHolder.from(R.drawable.mapbox_user_icon),
+  bearingImage = if (withBearing)
+    ImageHolder.from(R.drawable.mapbox_user_bearing_icon)
+  else
+    ImageHolder.from(R.drawable.mapbox_user_stroke_icon),
+  shadowImage = if (withBearing)
+    ImageHolder.from(R.drawable.mapbox_user_stroke_icon)
+  else
+    ImageHolder.from(R.drawable.mapbox_user_icon_shadow)
 )